Fix bug - '0' does not goto line start in vi mode.
The multiplier builder was eating the '0' keypress, preventing it from
being processed as the command to goto line beginning.

Added special handling to allow '0' to pass through the multiplier
builder if a multiplier has not yet been started.

'0' is still processed as a multiplier if a multiplier has been started
(that is, if a non-zero digit was entered before the 0).